+++ apps/mobile/src/components/common/icon.tsx
@@ -61,6 +61,7 @@ import PaletteBold from 'react-native-phosphor/src/bold/Palette'
 import PaperPlaneTiltBold from 'react-native-phosphor/src/bold/PaperPlaneTilt'
 import PauseBold from 'react-native-phosphor/src/bold/Pause'
 import PencilSimpleBold from 'react-native-phosphor/src/bold/PencilSimple'
+import PictureInPictureBold from 'react-native-phosphor/src/bold/PictureInPicture'
 import PlayBold from 'react-native-phosphor/src/bold/Play'
 import PlusBold from 'react-native-phosphor/src/bold/Plus'
 import PlusCircleBold from 'react-native-phosphor/src/bold/PlusCircle'
@@ -159,6 +160,7 @@ import PaletteDuotone from 'react-native-phosphor/src/duotone/Palette'
 import PaperPlaneTiltDuotone from 'react-native-phosphor/src/duotone/PaperPlaneTilt'
 import PauseDuotone from 'react-native-phosphor/src/duotone/Pause'
 import PencilSimpleDuotone from 'react-native-phosphor/src/duotone/PencilSimple'
+import PictureInPictureDuotone from 'react-native-phosphor/src/duotone/PictureInPicture'
 import PlayDuotone from 'react-native-phosphor/src/duotone/Play'
 import PlusDuotone from 'react-native-phosphor/src/duotone/Plus'
 import PlusCircleDuotone from 'react-native-phosphor/src/duotone/PlusCircle'
@@ -257,6 +259,7 @@ import PaletteFill from 'react-native-phosphor/src/fill/Palette'
 import PaperPlaneTiltFill from 'react-native-phosphor/src/fill/PaperPlaneTilt'
 import PauseFill from 'react-native-phosphor/src/fill/Pause'
 import PencilSimpleFill from 'react-native-phosphor/src/fill/PencilSimple'
+import PictureInPictureFill from 'react-native-phosphor/src/fill/PictureInPicture'
 import PlayFill from 'react-native-phosphor/src/fill/Play'
 import PlusFill from 'react-native-phosphor/src/fill/Plus'
 import PlusCircleFill from 'react-native-phosphor/src/fill/PlusCircle'
@@ -355,6 +358,7 @@ import PaletteRegular from 'react-native-phosphor/src/regular/Palette'
 import PaperPlaneTiltRegular from 'react-native-phosphor/src/regular/PaperPlaneTilt'
 import PauseRegular from 'react-native-phosphor/src/regular/Pause'
 import PencilSimpleRegular from 'react-native-phosphor/src/regular/PencilSimple'
+import PictureInPictureRegular from 'react-native-phosphor/src/regular/PictureInPicture'
 import PlayRegular from 'react-native-phosphor/src/regular/Play'
 import PlusRegular from 'react-native-phosphor/src/regular/Plus'
 import PlusCircleRegular from 'react-native-phosphor/src/regular/PlusCircle'
@@ -811,6 +815,12 @@ const icons = {
     duotone: PencilSimpleDuotone,
     fill: PencilSimpleFill,
     regular: PencilSimpleRegular,
+  },
+  PictureInPicture: {
+    bold: PictureInPictureBold,
+    duotone: PictureInPictureDuotone,
+    fill: PictureInPictureFill,
+    regular: PictureInPictureRegular,
   },
   Play: {
     bold: PlayBold,